Count seconds, follow the pointer, and say what it looks like
Three small things the engine's list asks for, and the first is a
correctness bug: a scrollbar told to hide after eighty frames hides
after two thirds of a second on one machine and after one and a third
on another. Everything timed here now takes seconds from `tick` - the
fade joins the caret blink, the double click and the momentum - and
`hide_after_frames` becomes `hide_after_seconds`. Ply's curve is
unchanged: full for the whole hold, then a fade a quarter as long
again. A program with no clock keeps its bars, which is the safer of
the two ways to be wrong.

`scrollHovered` moves whatever is under the pointer, innermost first
and one axis at a time, and says whether anything moved. `scrollBy`
wants a name, so a program with two lists had to redo the hit test it
had already asked for.

`cursor()` says what the pointer should look like, from a `.cursor` on
the element or a `setCursor` for the frame, with a text input asking
for a caret on its own. The names are Fluxion Platform's, so the
example's mapping is a switch rather than a table.

+++ src/text_input.zig
@@ -656,11 +656,11 @@ pub const TextEdit = struct {
     /// Whether the element was declared in the frame just finished. One that
     /// was not is dropped, exactly as a scroll position is.
     live: bool = false,
-    /// How many frames since the text or the cursor last moved, and whether
-    /// either did this frame. What a scrollbar told to hide itself counts -
-    /// the same pair a scroll container keeps, and folded in at the same
-    /// point in the frame.
-    idle: u32 = 0,
+    /// How long since the text or the cursor last moved, in seconds, and
+    /// whether either did this frame. What a scrollbar told to hide itself
+    /// reads - the same pair a scroll container keeps, advanced by the same
+    /// `Ui.tick`.
+    idle: f32 = 0,
     active: bool = false,
 
     pub const empty: TextEdit = .{};
